some ints (transaction id, cluster area) changed to long

+++ Clope/ClopeLib/Data/Cluster.cs
@@ -6,11 +6,11 @@ namespace ClopeLib.Data
 	public class Cluster : ICluster
 	{
 		// field
-		public int Area { get; private set; }
+		public long Area { get; private set; }
 
 		public bool IsEmpty => TransactionsCount == 0;
 
-		public int TransactionsCount { get; private set; }
+		public long TransactionsCount { get; private set; }
 
 		public int Width { get => counter.Positives; }
